Move and rename Edit to EditLink in components
